change: seperate files
- modal: remove bottomsheet
- bottomsheet: new file with several properties added
